dasdsain

+++ src/main/java/br/edu/impacta/model/Usuario.java
@@ -15,7 +15,7 @@ public class Usuario {
 	private String Nome;
 	private String Data_Nascimento;
 	private long Telefone;
-	private int CPF_CNPJ;
+	private long CPF_CNPJ;
 	private String Admin;
 	private String Email;
 	private String Senha;
@@ -57,10 +57,10 @@ public class Usuario {
 	public void setTelefone(long telefone) {
 		Telefone = telefone;
 	}
-	public int getCPF_CNPJ() {
+	public long getCPF_CNPJ() {
 		return CPF_CNPJ;
 	}
-	public void setCPF_CNPJ(int cPF_CNPJ) {
+	public void setCPF_CNPJ(long cPF_CNPJ) {
 		CPF_CNPJ = cPF_CNPJ;
 	}
 	public String getAdmin() {
